Threaded mips through.
The simplicity of changes like this reinforce the config maps approach.

The fact that the number of mips levels aren't part of an image, but most
functions taking an image also have to be given the number is unforgivable in a
modern language. The cost is so marginal and the ergonomics are game changing.