Prefer `tap` over `send!`.
The issue is that during the transduction phase, values returned via emit move
on, but those returned to the caller are dropped. It's a form of CPS.

What I don't understand is how the returned values make it to the reducer in the
case of `send!`. They should be dropped just as in the transduction phase, but
are not.

Maybe it's an error in my impl of reduction over channels...