Added fast non persistent `into!` function.
This lets us dump the product of transducers into Base.Vector in times
comparable to Base.Iterators.

If you annotate the expected type (by passing a typed empty vector) to `into!`,
you will get the best performance. If you don't know the eltype, `into!` will
try to deduce it (using yet another transducer!).

This inferring transducer has some overhead, but the savings vs boxing are
enormous, so it's well worth it most of the time. The optional `infertype` arg
lets you turn it off if you know it's useless.