Converting C style enums to symbols.
Separates the concerns of defining pipelines from linking to the C library
nicely, but we lose static checking of valid flags.

Well placed assertions have so far made debugging not too painful, but excessive
dynamism has downsides.

The real test of this project will be how well I mitigate that.