Ported viking test to new pipelines and cleaned up.
I'm trying to decide if binding externally allocated descriptors outside of the
graphics pipeline is an ugly hack or the right thing to do in general.

When appropriate, the method used in viking.jl is the right thing to do. Why
reallocate and rebind unnecessarily?

But I'm dubious that this will be the solution I want when passing data between
shaders.